+++ proxies/collection/permission.js
@@ -1,0 +1,13 @@
+import { Proxies } from './';
+
+Proxies.allow({
+  insert: function () {
+    return true;
+  },
+  update: function () {
+    return false;
+  },
+  remove: function () {
+    return false;
+  }
+});
